⚡️ perf(home): cut the first navigation's closure in half
The web main layout statically mounted every Electron-only feature behind
isDesktop && (title bar, tab bridges, overlays, zoom HUD); index.desktop.tsx
already owns them, so the web layout keeps only the auth recovery pair. The
title bar's update notification alone dragged the Markdown stack (elkjs,
katex, octicons, mermaid, parse5) into every web route.

Modal bodies and markdown views that the home route only opens on demand
now mount through lazy(): the inbox rail, notification detail, news and
unread-topic previews, the brief summary and its comment editor, the
changelog, create-project (pinyin-pro), task template detail, the
error-fallback highlighter, the drag preview's file icon, and the
agent-settings opener that useInterceptingRoutes pulled into the hotkey
registration. The boot-path eslint rule now also covers the main layout,
the home route, HomeSidebar and NavPanel.

Web production build, home route closure (main layout + home page, chunk
graph): 200 chunks / 2103 KB gzip -> 182 / 1134 KB.
