+++ src/components/ErrorBoundary/AlertFallback.tsx
@@ -1,8 +1,9 @@
 'use client';
 
-import { Highlighter } from '@lobehub/ui';
 import { Alert } from '@lobehub/ui/base-ui';
-import { memo } from 'react';
+import { lazy, memo, Suspense } from 'react';
+
+const Highlighter = lazy(() => import('@lobehub/ui/es/Highlighter/index'));
 
 interface AlertFallbackProps {
   error: Error;
@@ -22,9 +23,16 @@ const AlertFallback = memo<AlertFallbackProps>(({ error, resetErrorBoundary, tit
       type="secondary"
       extra={
         error?.stack ? (
-          <Highlighter actionIconSize="small" language="plaintext" padding={8} variant="borderless">
-            {error.stack}
-          </Highlighter>
+          <Suspense fallback={null}>
+            <Highlighter
+              actionIconSize="small"
+              language="plaintext"
+              padding={8}
+              variant="borderless"
+            >
+              {error.stack}
+            </Highlighter>
+          </Suspense>
         ) : undefined
       }
       onClose={resetErrorBoundary}
