+++ src/features/DailyBrief/__tests__/BriefCardSummary.test.tsx
@@ -21,10 +21,12 @@ vi.mock('react-i18next', () => ({
 }));
 
 describe('BriefCardSummary', () => {
-  it('should render summary text', () => {
+  it('should render summary text', async () => {
     vi.mocked(useSize).mockReturnValue(undefined);
     render(<BriefCardSummary summary="Test summary content" />);
-    expect(screen.getByText('Test summary content')).toBeInTheDocument();
+    expect(
+      await screen.findByText('Test summary content', {}, { timeout: 10_000 }),
+    ).toBeInTheDocument();
   });
 
   it('should not show expand link when content does not overflow', () => {
