+++ src/routes/(main)/_layout/index.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { HotkeyScopeEnum } from '@lobechat/const/hotkeys';
-import { TITLE_BAR_HEIGHT } from '@lobechat/desktop-bridge';
 import { Flexbox } from '@lobehub/ui';
 import { cx } from 'antd-style';
 import { type FC } from 'react';
@@ -13,17 +12,8 @@ import WorkspaceContextSlot from '@/business/client/WorkspaceContextSlot';
 import RouteSegmentSkeleton from '@/components/Skeleton/RouteSegment';
 import { isDesktop } from '@/const/version';
 import { BANNER_HEIGHT } from '@/features/AlertBanner/CloudBanner';
-import DesktopBrowserGatewayBridge from '@/features/DesktopBrowserGatewayBridge';
-import DesktopFileMenuBridge from '@/features/DesktopFileMenuBridge';
 import DesktopLayoutContainer from '@/features/DesktopLayoutContainer';
-import DesktopNavigationBridge from '@/features/DesktopNavigationBridge';
 import AuthRequiredModal from '@/features/Electron/AuthRequiredModal';
-import OverlayCaptureUploader from '@/features/Electron/ScreenCapture/OverlayCaptureUploader';
-import OverlayMessageDispatcher from '@/features/Electron/ScreenCapture/OverlayMessageDispatcher';
-import OverlaySnapshotPublisher from '@/features/Electron/ScreenCapture/OverlaySnapshotPublisher';
-import ZoomHUD from '@/features/Electron/system/ZoomHUD';
-import TabCacheBridges from '@/features/Electron/titlebar/TabBar/TabCacheBridges';
-import TitleBar from '@/features/Electron/titlebar/TitleBar';
 import HotkeyHelperPanel from '@/features/HotkeyHelperPanel';
 import NavPanelShell from '@/features/NavPanel/Shell';
 import { DndContextWrapper } from '@/features/ResourceManager/DndContextWrapper';
@@ -42,6 +32,9 @@ import { styles } from './style';
 const CloudBanner = dynamic(() => import('@/features/AlertBanner/CloudBanner'));
 const GlobalApprovalNotification = dynamic(() => import('@/features/GlobalApprovalNotification'));
 
+// The Electron shell (title bar, tab bridges, overlays, zoom HUD) lives in
+// index.desktop.tsx; only the auth recovery pair stays here so a desktop build
+// that falls back to this layout can still re-login.
 const Layout: FC = () => {
   const { isPWA } = usePlatform();
   const { showCloudPromotion } = useServerConfigStore(featureFlagsSelectors);
@@ -52,31 +45,13 @@ const Layout: FC = () => {
       {isDesktop && <AuthRequiredModal />}
       <WorkspaceContextSlot>
         <RouteMetaBridge />
-        {isDesktop && <TabCacheBridges />}
-        <Suspense fallback={null}>
-          {isDesktop && <DesktopNavigationBridge />}
-          {isDesktop && <DesktopFileMenuBridge />}
-          {isDesktop && <DesktopBrowserGatewayBridge />}
-          {isDesktop && <OverlaySnapshotPublisher />}
-          {isDesktop && <OverlayCaptureUploader />}
-          {isDesktop && <OverlayMessageDispatcher />}
-          {showCloudPromotion && <CloudBanner />}
-        </Suspense>
-        {isDesktop && <ZoomHUD />}
-
-        <Suspense fallback={null}>{isDesktop && <TitleBar />}</Suspense>
+        <Suspense fallback={null}>{showCloudPromotion && <CloudBanner />}</Suspense>
         <DndContextWrapper>
           <Flexbox
             horizontal
             className={cx(isPWA ? styles.mainContainerPWA : styles.mainContainer)}
+            height={showCloudPromotion ? `calc(100% - ${BANNER_HEIGHT}px)` : '100%'}
             width={'100%'}
-            height={
-              isDesktop
-                ? `calc(100% - ${TITLE_BAR_HEIGHT}px)`
-                : showCloudPromotion
-                  ? `calc(100% - ${BANNER_HEIGHT}px)`
-                  : '100%'
-            }
           >
             <NavPanelShell />
             <DesktopLayoutContainer>
